+++ src/isolate/generic/handle_cast.h
@@ -1,6 +1,7 @@
 #pragma once
 #include <v8.h>
 #include "error.h"
+#include "../v8_version.h"
 
 namespace ivm {
 
@@ -231,6 +232,20 @@ inline auto HandleCastImpl(int32_t value, const HandleCastArguments& arguments, 
 
 inline auto HandleCastImpl(uint32_t value, const HandleCastArguments& arguments, HandleCastTag<v8::Local<v8::Integer>> /*tag*/) {
 	return v8::Integer::NewFromUnsigned(arguments.isolate, value);
+}
+
+#if V8_AT_LEAST(6, 9, 258)
+inline auto HandleCastImpl(int64_t value, const HandleCastArguments& arguments, HandleCastTag<v8::Local<v8::BigInt>> /*tag*/) {
+	return v8::BigInt::New(arguments.isolate, value);
+}
+
+inline auto HandleCastImpl(uint64_t value, const HandleCastArguments& arguments, HandleCastTag<v8::Local<v8::BigInt>> /*tag*/) {
+	return v8::BigInt::NewFromUnsigned(arguments.isolate, value);
+}
+#endif
+
+inline auto HandleCastImpl(double value, const HandleCastArguments& arguments, HandleCastTag<v8::Local<v8::Number>> /*tag*/) {
+	return v8::Number::New(arguments.isolate, value);
 }
 
 inline auto HandleCastImpl(const char* value, const HandleCastArguments& arguments, HandleCastTag<v8::Local<v8::String>> /*tag*/) {