+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 - `result` property on `eval` and `evalClosure` has been removed. The result is now just the return
 value.
 - All `isolated-vm` class prototypes, and most instances are frozen.
+- `isolate.cpuTime` and `isolate.wallTime` now return bigints.